+++ src/commands/copySymbolToClipboard.ts
@@ -1,10 +1,11 @@
-import { env, ProgressLocation, TextEditor, window } from 'vscode';
+import { CancellationToken, CancellationTokenSource, env, ProgressLocation, TextEditor, window } from 'vscode';
 import {
   SymbolInformation,
   TextDocumentPositionParams,
 } from 'vscode-languageclient/node';
 import { SorbetExtensionContext } from '../sorbetExtensionContext';
 import { ServerStatus } from '../types';
+import { Log } from '../log';
 
 /**
  * Copy symbol at current.
@@ -15,9 +16,7 @@ export async function copySymbolToClipboard(
   editor: TextEditor,
 ): Promise<void> {
   if (!editor.selection.isEmpty) {
-    context.log.debug(
-      'CopySymbol: Non-empty selection, cannot determine target symbol.',
-    );
+    context.log.warn('CopySymbol: Non-empty selection, cannot determine target symbol.');
     return;
   }
 
@@ -42,32 +41,54 @@ export async function copySymbolToClipboard(
     position,
   };
 
-  let canceled = false;
-  const symbolInfo = await window.withProgress(
-    {
-      cancellable: true,
-      location: ProgressLocation.Notification,
-    },
-    async (progress, token) => {
-      progress.report({ message: 'Querying Sorbet …' });
-      const r = await client.sendRequest<SymbolInformation>(
-        'sorbet/showSymbol',
-        params,
-        token,
-      );
-
-      canceled = token.isCancellationRequested;
-      return r || undefined;
-    },
-  );
+  // If Sorbet is busy, retrieving symbol information might take a while. To
+  // avoid having long operation surprisingly overwrite the current clipboard
+  // contents, a cancelable progress notification is shown after 2s.
+  const symbolInfo = await withProgress(
+    (token) => client.sendRequest<SymbolInformation>('sorbet/showSymbol', params, token),
+    2000,
+    context.log);
 
   if (symbolInfo) {
     await env.clipboard.writeText(symbolInfo.name);
-    context.log.debug('CopySymbol: Copied symbol name', symbolInfo.name);
-  } else if (!canceled) {
-    window.showWarningMessage('No symbol information found current location');
-    context.log.debug('CopySymbol: No symbol name');
+    context.log.info('CopySymbol: Copied symbol to clipboard:', symbolInfo.name);
   } else {
-    context.log.debug('CopySymbol: Canceled operation');
+    context.log.info('CopySymbol: No symbol found.');
   }
+}
+
+async function withProgress<T>(task: (token: CancellationToken) => Promise<T>, progressDelayMS: number, log: Log): Promise<T | undefined> {
+  let resolved = false;
+  const cts = new CancellationTokenSource();
+  const resultPromise = task(cts.token)
+    .then((r) => {
+      log.trace('CopySymbol: Operation completed:', r);
+      resolved = true;
+      return cts.token.isCancellationRequested ? undefined : r;
+    });
+
+  await Promise.race([
+    new Promise<void>(resolve => setTimeout(resolve, progressDelayMS))
+      .then(() => {
+        if (!resolved) {
+          window.withProgress(
+            {
+              location: ProgressLocation.Notification,
+              title: 'Querying Sorbet …',
+              cancellable: true,
+            },
+            (_, token) => {
+              token.onCancellationRequested(() => {
+                log.trace('CopySymbol: Operation canceled.');
+                cts.cancel();
+              });
+              return resultPromise;
+            },
+          );
+        }
+      }),
+    resultPromise,
+  ]);
+
+  return resultPromise;
 }